CNN_MODEL/requirements.txt
certifi==2021.10.8
click==8.0.3
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==2.0.2
wincertstore==0.2
pillow==8.4.0
flask-cors
tensorflow==2.9.1

